1. 修改config类 2. 增加db和redis操作类(待修改，无法编译通过) 3. 增加Application类，封装Server和Client

+++ net/client/ClientLoop.cc
@@ -34,7 +34,7 @@ namespace MF {
             return client;
         }
 
-        const std::map<uint32_t , std::shared_ptr<MyClient>>& ClientLoop::getClientMap() const {
+        const std::map<uint64_t , std::shared_ptr<MyClient>>& ClientLoop::getClientMap() const {
             return clients;
         }
